requirements.txt
autopep8==1.6.0
certifi==2021.10.8
distlib==0.3.4
dlib==19.23.0
dnspython==2.2.0
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.4.2
Flask==2.0.3
greenlet==1.1.2
imutils==0.5.4
itsdangerous==2.1.1
Jinja2==3.0.3
latexcodec==2.0.1
MarkupSafe==2.1.0
msgpack==1.0.3
numpy==1.22.3
omf==1.0.1
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
packaging==20.9
pagelabels==1.2.0
pandas==1.4.1
pdfrw==0.4
Pillow==9.0.1
pkgconfig==1.5.5
platformdirs==2.4.1
properties==0.6.1
psutil==5.9.0
pybtex==0.24.0
pycodestyle==2.8.0
PyMuPDF==1.19.4
pynvim==0.4.3
pyparsing==3.0.6
pyperclip==1.8.2
pypng==0.0.21
python-dateutil==2.8.2
pytz==2021.3
roman==3.3
six==1.16.0
toml==0.10.2
vectormath==0.2.2
virtualenv==20.13.0
virtualfish==2.5.4
Werkzeug==2.0.3
